@@ -4,66 +4,39 @@ import (
 	"encoding/json"
 	"errors"
 	"github.com/RBG-TUM/commons"
-	"github.com/gabstv/melody"
 	"github.com/getsentry/sentry-go"
 	"github.com/gin-gonic/gin"
 	"github.com/joschahenningsen/TUM-Live/dao"
 	"github.com/joschahenningsen/TUM-Live/model"
 	"github.com/joschahenningsen/TUM-Live/tools"
+	"github.com/joschahenningsen/TUM-Live/tools/realtime"
 	"github.com/joschahenningsen/TUM-Live/tools/tum"
 	log "github.com/sirupsen/logrus"
 	"sync"
 )
 
 const (
+	LiveUpdateRoomName       = "live-update"
 	UpdateTypeCourseWentLive = "course_went_live"
 )
-
-var liveMelody *melody.Melody
 
 var liveUpdateListenerMutex sync.RWMutex
 var liveUpdateListener = map[uint]*liveUpdateUserSessionsWrapper{}
 
-const maxLiveUpdateParticipants = 10000
-
-type liveUpdateRoutes struct {
-	dao.DaoWrapper
-}
-
 type liveUpdateUserSessionsWrapper struct {
-	sessions []*melody.Session
+	sessions []*realtime.Context
 	courses  []uint
 }
 
-func configGinLiveUpdateRouter(router *gin.RouterGroup, daoWrapper dao.DaoWrapper) {
-	routes := liveUpdateRoutes{daoWrapper}
-
-	router.GET("/ws", routes.handleLiveConnect)
-
-	if liveMelody == nil {
-		log.Printf("creating liveMelody")
-		liveMelody = melody.New()
-	}
-
-	liveMelody.HandleConnect(liveUpdateConnectionHandler)
-	liveMelody.HandleDisconnect(liveUpdateDisconnectHandler)
+func RegisterLiveUpdateRealtimeChannel() {
+	RealtimeInstance.RegisterChannel(LiveUpdateRoomName, realtime.ChannelHandlers{
+		OnSubscribe:   liveUpdateOnSubscribe,
+		OnUnsubscribe: liveUpdateOnUnsubscribe,
+	})
 }
 
-func (r liveUpdateRoutes) handleLiveConnect(c *gin.Context) {
-	// max participants in chat to prevent attacks
-	if liveMelody.Len() > maxLiveUpdateParticipants {
-		return
-	}
-
-	ctxMap := make(map[string]interface{}, 1)
-	ctxMap["ctx"] = c
-	ctxMap["dao"] = r.DaoWrapper
-
-	_ = liveMelody.HandleRequestWithKeys(c.Writer, c.Request, ctxMap)
-}
-
-func liveUpdateDisconnectHandler(s *melody.Session) {
-	ctx, _ := s.Get("ctx") // get gin context
+func liveUpdateOnUnsubscribe(psc *realtime.Context) {
+	ctx, _ := psc.Client.Get("ctx") // get gin context
 	foundContext, exists := ctx.(*gin.Context).Get("TUMLiveContext")
 	if !exists {
 		sentry.CaptureException(errors.New("context should exist but doesn't"))
@@ -78,9 +51,9 @@ func liveUpdateDisconnectHandler(s *melody.Session) {
 	}
 
 	liveUpdateListenerMutex.Lock()
-	var newSessions []*melody.Session
+	var newSessions []*realtime.Context
 	for _, session := range liveUpdateListener[userId].sessions {
-		if session != s {
+		if session != psc {
 			newSessions = append(newSessions, session)
 		}
 	}
@@ -92,9 +65,9 @@ func liveUpdateDisconnectHandler(s *melody.Session) {
 	liveUpdateListenerMutex.Unlock()
 }
 
-var liveUpdateConnectionHandler = func(s *melody.Session) {
-	ctx, _ := s.Get("ctx") // get gin context
-	daoWrapper, _ := s.Get("dao")
+func liveUpdateOnSubscribe(psc *realtime.Context) {
+	ctx, _ := psc.Client.Get("ctx") // get gin context
+	daoWrapper, _ := psc.Client.Get("dao")
 
 	foundContext, exists := ctx.(*gin.Context).Get("TUMLiveContext")
 	if !exists {
@@ -130,9 +103,9 @@ var liveUpdateConnectionHandler = func(s *melody.Session) {
 
 	liveUpdateListenerMutex.Lock()
 	if liveUpdateListener[userId] != nil {
-		liveUpdateListener[userId] = &liveUpdateUserSessionsWrapper{append(liveUpdateListener[userId].sessions, s), courses}
+		liveUpdateListener[userId] = &liveUpdateUserSessionsWrapper{append(liveUpdateListener[userId].sessions, psc), courses}
 	} else {
-		liveUpdateListener[userId] = &liveUpdateUserSessionsWrapper{[]*melody.Session{s}, courses}
+		liveUpdateListener[userId] = &liveUpdateUserSessionsWrapper{[]*realtime.Context{psc}, courses}
 	}
 	liveUpdateListenerMutex.Unlock()
 }
@@ -144,7 +117,7 @@ func NotifyLiveUpdateCourseWentLive(courseId uint) {
 		for _, course := range userWrap.courses {
 			if course == courseId {
 				for _, session := range userWrap.sessions {
-					_ = session.Write(updateMessage)
+					_ = session.Send(updateMessage)
 				}
 				break
 			}
