@@ -40,11 +40,12 @@ func GinServer() (err error) {
 
 	router.Use(tools.InitContext(dao.NewDaoWrapper()))
 
+	liveUpdates := router.Group("/api/pub-sub")
+	api.ConfigRealtimeRouter(liveUpdates)
+
 	// event streams don't work with gzip, configure group without
 	chat := router.Group("/api/chat")
 	api.ConfigChatRouter(chat)
-	liveUpdates := router.Group("/api/live-update")
-	api.ConfigLiveUpdateRouter(liveUpdates)
 
 	router.Use(gzip.Gzip(gzip.DefaultCompression))
 	api.ConfigGinRouter(router)
